refactor: move mock data and wrapper to mocks file
